record state root hash to resolve statedb smt rebuild after server restart
