+++ crates/rooch-executor/src/actor/executor.rs
@@ -28,9 +28,6 @@ use moveos_types::transaction::FunctionCall;
 use moveos_types::transaction::TransactionExecutionInfo;
 use moveos_types::transaction::VerifiedMoveOSTransaction;
 use moveos_types::tx_context::TxContext;
-// use raw_store::rocks::RocksDB;
-// use raw_store::StoreInstance;
-// use rooch_config::store_config::StoreConfig;
 use rooch_framework::bindings::address_mapping::AddressMapping;
 use rooch_framework::bindings::auth_validator::AuthValidatorCaller;
 use rooch_framework::bindings::transaction_validator::TransactionValidator;
