+++ crates/rooch/src/commands/move_cli/commands/unit_test.rs
@@ -11,7 +11,7 @@ use move_package::BuildConfig;
 use move_unit_test::extensions::set_extension_hook;
 use move_vm_runtime::native_extensions::NativeContextExtensions;
 use moveos_stdlib::natives::moveos_stdlib::raw_table::NativeTableContext;
-use moveos_store::state_store::StateDBStore;
+use moveos_store::state_store::statedb::StateDBStore;
 use moveos_store::MoveOSStore;
 use moveos_verifier::build::build_model_with_test_attr;
 use moveos_verifier::metadata::run_extended_checks;
@@ -88,7 +88,7 @@ impl Test {
 }
 
 static STATEDBSTORE: Lazy<Box<StateDBStore>> =
-    Lazy::new(|| Box::new(MoveOSStore::mock().unwrap().state_store));
+    Lazy::new(|| Box::new(MoveOSStore::mock_moveos_store().unwrap().statedb));
 
 fn new_moveos_natives_runtime(ext: &mut NativeContextExtensions) {
     let statedb_store = Lazy::force(&STATEDBSTORE).as_ref();
