+++ moveos/moveos-store/Cargo.toml
@@ -12,8 +12,6 @@ repository = { workspace = true }
 rust-version = { workspace = true }
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-#[lib]
-#proc-macro = true
 
 [dependencies]
 anyhow = { workspace = true }
@@ -23,7 +21,6 @@ serde = { workspace = true }
 serde_bytes = { workspace = true }
 hex = { workspace = true }
 parking_lot = { workspace = true }
-#proc-macro2 = { workspace = true }
 quote = { workspace = true }
 num_enum = { workspace = true }
 once_cell = { workspace = true }
