+++ moveos/moveos-store/src/lib.rs
@@ -4,15 +4,17 @@
 // Copyright (c) The Starcoin Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use anyhow::Result;
+use anyhow::{Error, Result};
 use once_cell::sync::Lazy;
 use raw_store::{ColumnFamilyName, StoreInstance};
 use std::collections::BTreeMap;
 use std::fmt::{Debug, Display, Formatter};
 use std::sync::Arc;
 
+use crate::config_store::{ConfigStore, StartupInfoDBStore};
 use crate::event_store::{EventDBStore, EventStore};
-use crate::state_store::StateDBStore;
+use crate::state_store::statedb::StateDBStore;
+use crate::state_store::NodeDBStore;
 use crate::transaction_store::{TransactionDBStore, TransactionStore};
 use move_core_types::language_storage::TypeTag;
 use moveos_config::store_config::RocksdbConfig;
@@ -20,10 +22,14 @@ use moveos_types::event::{Event, EventID};
 use moveos_types::event_filter::EventFilter;
 use moveos_types::h256::H256;
 use moveos_types::object::ObjectID;
+use moveos_types::startup_info::StartupInfo;
+use moveos_types::state::State;
+use moveos_types::state_resolver::StateResolver;
 use moveos_types::transaction::TransactionExecutionInfo;
 use raw_store::rocks::RocksDB;
 use smt::NodeStore;
 
+pub mod config_store;
 pub mod event_store;
 pub mod state_store;
 #[cfg(test)]
@@ -35,6 +41,7 @@ pub const STATE_NODE_PREFIX_NAME: ColumnFamilyName = "state_node";
 pub const TRANSACTION_PREFIX_NAME: ColumnFamilyName = "transaction";
 pub const EVENT_PREFIX_NAME: ColumnFamilyName = "event";
 pub const EVENT_INDEX_PREFIX_NAME: ColumnFamilyName = "event_index";
+pub const CONFIG_PREFIX_NAME: ColumnFamilyName = "config";
 
 ///db store use prefix_name vec to init
 /// Please note that adding a prefix needs to be added in vec simultaneously, remember！！
@@ -44,6 +51,7 @@ static VEC_PREFIX_NAME: Lazy<Vec<ColumnFamilyName>> = Lazy::new(|| {
         TRANSACTION_PREFIX_NAME,
         EVENT_PREFIX_NAME,
         EVENT_INDEX_PREFIX_NAME,
+        CONFIG_PREFIX_NAME,
     ]
 });
 
@@ -56,18 +64,17 @@ impl StoreMeta {
     }
 }
 
-// #[derive(Clone)]
-pub struct MoveOSStore {
-    pub state_store: StateDBStore,
+pub struct MoveOSDB {
+    pub node_store: NodeDBStore,
     pub event_store: EventDBStore,
     pub transaction_store: TransactionDBStore,
+    pub config_store: StartupInfoDBStore,
 }
-// // TODO: remove Arc<dyn Store>, we can clone Store directly.
 
-impl MoveOSStore {
-    pub fn mock() -> Result<Self> {
+impl MoveOSDB {
+    pub fn mock_store_instance() -> StoreInstance {
         let tmpdir = moveos_config::temp_dir();
-        let mock_instance = StoreInstance::new_db_instance(
+        StoreInstance::new_db_instance(
             RocksDB::new(
                 tmpdir.path(),
                 StoreMeta::get_column_family_names().to_vec(),
@@ -75,40 +82,78 @@ impl MoveOSStore {
                 None,
             )
             .unwrap(),
-        );
+        )
+    }
 
-        Self::new(mock_instance)
+    pub fn mock_moveosdb() -> Result<Self> {
+        Self::new(Self::mock_store_instance())
     }
 
     pub fn new(instance: StoreInstance) -> Result<Self> {
         let store = Self {
-            state_store: StateDBStore::new(instance.clone()),
+            node_store: NodeDBStore::new(instance.clone()),
             event_store: EventDBStore::new(instance.clone()),
-            transaction_store: TransactionDBStore::new(instance),
+            transaction_store: TransactionDBStore::new(instance.clone()),
+            config_store: StartupInfoDBStore::new(instance),
+        };
+        Ok(store)
+    }
+}
+
+pub struct MoveOSStore {
+    pub statedb: StateDBStore,
+    pub moveosdb: MoveOSDB,
+}
+
+impl MoveOSStore {
+    pub fn mock_moveos_store() -> Result<Self> {
+        let moveosdb = MoveOSDB::mock_moveosdb()?;
+        Self::new(moveosdb)
+    }
+
+    pub fn new(moveosdb: MoveOSDB) -> Result<Self> {
+        let store = Self {
+            statedb: StateDBStore::new(moveosdb.node_store.clone()),
+            moveosdb,
+        };
+        Ok(store)
+    }
+
+    pub fn new_with_root(moveosdb: MoveOSDB, state_root: Option<H256>) -> Result<Self> {
+        let store = Self {
+            statedb: StateDBStore::new_with_root(moveosdb.node_store.clone(), state_root),
+            moveosdb,
         };
         Ok(store)
     }
 
     pub fn get_event_store(&self) -> &EventDBStore {
-        &self.event_store
+        &self.moveosdb.event_store
     }
 
     pub fn get_transaction_store(&self) -> &TransactionDBStore {
-        &self.transaction_store
+        &self.moveosdb.transaction_store
+    }
+
+    pub fn get_state_node_store(&self) -> &NodeDBStore {
+        &self.moveosdb.node_store
+    }
+
+    pub fn get_config_store(&self) -> &StartupInfoDBStore {
+        &self.moveosdb.config_store
     }
 
     pub fn get_state_store(&self) -> &StateDBStore {
-        &self.state_store
+        &self.statedb
     }
 }
 
 impl Display for MoveOSStore {
     fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-        // write!(f, "{}", self.clone())
-
-        write!(f, "state_store")?;
+        write!(f, "statedb")?;
         write!(f, "event_store")?;
         write!(f, "transaction_store")?;
+        write!(f, "node_store")?;
         Ok(())
     }
 }
@@ -121,33 +166,33 @@ impl Debug for MoveOSStore {
 
 impl NodeStore for MoveOSStore {
     fn get(&self, hash: &H256) -> Result<Option<Vec<u8>>> {
-        self.state_store.node_store.get(hash)
+        self.get_state_node_store().get(hash)
     }
 
     fn put(&self, key: H256, node: Vec<u8>) -> Result<()> {
-        self.state_store.node_store.put(key, node)
+        self.get_state_node_store().put(key, node)
     }
 
     fn write_nodes(&self, nodes: BTreeMap<H256, Vec<u8>>) -> Result<()> {
-        self.state_store.node_store.write_nodes(nodes)
+        self.get_state_node_store().write_nodes(nodes)
     }
 }
 
 impl EventStore for MoveOSStore {
     fn save_event(&self, event: Event) -> Result<()> {
-        self.event_store.save_event(event)
+        self.get_event_store().save_event(event)
     }
 
     fn save_events(&self, events: Vec<Event>) -> Result<()> {
-        self.event_store.save_events(events)
+        self.get_event_store().save_events(events)
     }
 
     fn get_event(&self, event_id: EventID) -> Result<Option<Event>> {
-        self.event_store.get_event(event_id)
+        self.get_event_store().get_event(event_id)
     }
 
     fn get_events_by_tx_hash(&self, tx_hash: &H256) -> Result<Vec<Event>> {
-        self.event_store.get_events_by_tx_hash(tx_hash)
+        self.get_event_store().get_events_by_tx_hash(tx_hash)
     }
 
     fn get_events_by_event_handle_id(
@@ -156,39 +201,53 @@ impl EventStore for MoveOSStore {
         cursor: Option<u64>,
         limit: u64,
     ) -> Result<Vec<Event>> {
-        self.event_store
+        self.get_event_store()
             .get_events_by_event_handle_id(event_handle_id, cursor, limit)
     }
 
     fn get_events_by_event_handle_type(&self, event_handle_type: &TypeTag) -> Result<Vec<Event>> {
-        self.event_store
+        self.get_event_store()
             .get_events_by_event_handle_type(event_handle_type)
     }
 
     fn get_events_with_filter(&self, filter: EventFilter) -> Result<Vec<Event>> {
-        self.event_store.get_events_with_filter(filter)
+        self.get_event_store().get_events_with_filter(filter)
     }
 }
 
 impl TransactionStore for MoveOSStore {
     fn save_tx_exec_info(&self, tx_exec_info: TransactionExecutionInfo) -> Result<()> {
-        self.transaction_store.save_tx_exec_info(tx_exec_info)
+        self.get_transaction_store().save_tx_exec_info(tx_exec_info)
     }
 
     fn get_tx_exec_info(&self, tx_hash: H256) -> Result<Option<TransactionExecutionInfo>> {
-        self.transaction_store.get_tx_exec_info(tx_hash)
+        self.get_transaction_store().get_tx_exec_info(tx_hash)
     }
 
     fn multi_get_tx_exec_infos(
         &self,
         tx_hashes: Vec<H256>,
     ) -> Result<Vec<Option<TransactionExecutionInfo>>> {
-        self.transaction_store.multi_get_tx_exec_infos(tx_hashes)
+        self.get_transaction_store()
+            .multi_get_tx_exec_infos(tx_hashes)
+    }
+}
+
+impl ConfigStore for MoveOSStore {
+    fn get_startup_info(&self) -> Result<Option<StartupInfo>> {
+        self.get_config_store().get_startup_info()
+    }
+
+    fn save_startup_info(&self, startup_info: StartupInfo) -> Result<()> {
+        self.get_config_store().save_startup_info(startup_info)
     }
 }
 
 /// Moveos store define
-pub trait Store: NodeStore + TransactionStore + EventStore + IntoSuper<dyn NodeStore> {}
+pub trait Store:
+    NodeStore + TransactionStore + EventStore + ConfigStore + IntoSuper<dyn NodeStore>
+{
+}
 
 pub trait IntoSuper<Super: ?Sized> {
     fn as_super(&self) -> &Super;
@@ -213,3 +272,13 @@ impl<'a, T: 'a + NodeStore> IntoSuper<dyn NodeStore + 'a> for T {
 }
 
 impl Store for MoveOSStore {}
+
+impl StateResolver for MoveOSStore {
+    fn resolve_state(
+        &self,
+        handle: &ObjectID,
+        key: &[u8],
+    ) -> std::result::Result<Option<State>, Error> {
+        self.statedb.resolve_state(handle, key)
+    }
+}
