+++ moveos/moveos-store/src/tests/mod.rs
@@ -1,4 +1,5 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
+mod test_state_store;
 mod test_store;
