+++ moveos/moveos-store/src/tests/test_state_store.rs
@@ -1,16 +1,19 @@
 // Copyright (c) RoochNetwork
 // SPDX-License-Identifier: Apache-2.0
 
+use move_core_types::effects::{ChangeSet, Op};
+use moveos_types::h256::H256;
+use moveos_types::object::ObjectID;
+use moveos_types::state::{MoveState, StateChangeSet};
+use moveos_types::{move_string::MoveString, state::TableChange};
+use smt::NodeStore;
 use std::str::FromStr;
 
-use moveos_types::{move_string::MoveString, state::TableChange};
-
-use super::*;
 use crate::MoveOSStore;
 
 #[test]
 fn test_statedb() {
-    let moveos_store = MoveOSStore::mock().unwrap();
+    let moveos_store = MoveOSStore::mock_moveos_store().unwrap();
 
     let change_set = ChangeSet::new();
 
@@ -26,7 +29,7 @@ fn test_statedb() {
 
     table_change_set.changes.insert(table_handle, table_change);
     moveos_store
-        .state_store
+        .get_state_store()
         .apply_change_set(change_set, table_change_set)
         .unwrap();
 
@@ -36,4 +39,23 @@ fn test_statedb() {
         .unwrap();
     assert!(state.is_some());
     assert_eq!(state.unwrap(), value.into());
+}
+
+#[test]
+fn test_reopen() {
+    let moveos_store = MoveOSStore::mock_moveos_store().unwrap();
+    let node_store = moveos_store.get_state_node_store();
+
+    let key = H256::random();
+    let node = b"testnode".to_vec();
+    {
+        node_store
+            .put(key, node.clone())
+            .map_err(|e| anyhow::anyhow!("test_state_store test_reopen error: {:?}", e))
+            .ok();
+        assert_eq!(node_store.get(&key).unwrap(), Some(node.clone()));
+    }
+    {
+        assert_eq!(node_store.get(&key).unwrap(), Some(node));
+    }
 }
