+++ moveos/moveos-types/src/lib.rs
@@ -17,6 +17,7 @@ pub mod move_string;
 pub mod move_types;
 pub mod object;
 pub mod serde;
+pub mod startup_info;
 pub mod state;
 pub mod state_resolver;
 pub mod storage_context;
